@@ -10,10 +10,11 @@
     "logs": "firebase functions:log"
   },
   "engines": {
-    "node": "8"
+    "node": "10"
   },
   "main": "lib/index.js",
   "dependencies": {
+    "algoliasearch": "^4.3.1",
     "firebase-admin": "^8.10.0",
     "firebase-functions": "^3.6.1"
   },
