@@ -1,6 +1,7 @@
-import { BrowserModule } from '@angular/platform-browser';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-
+import {BrowserModule} from '@angular/platform-browser';
+import {FormsModule, ReactiveFormsModule} from '@angular/forms';
+import {HashLocationStrategy, LocationStrategy} from '@angular/common'
+import {MatChipsModule} from '@angular/material/chips';
 import {AngularFireAuthModule} from '@angular/fire/auth';
 import {AngularFireDatabaseModule} from '@angular/fire/database';
 import {AngularFireModule} from '@angular/fire';
@@ -8,15 +9,24 @@ import {AngularFirestoreModule} from '@angular/fire/firestore';
 import {AngularFireStorageModule} from '@angular/fire/storage';
 import {AppComponent} from './app.component';
 import {AppRoutingModule} from './app-routing.module';
+import {AutoCompleteComponent} from './auto-complete/auto-complete.component';
 import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
+import {ChangeProfileImgComponent} from './change-profile-img/change-profile-img.component';
 import {CurrentProfilePageComponent} from './current-profile-page/current-profile-page.component';
 import {EmailComponent} from './email/email.component';
+import {environment} from '../environments/environment';
+import {GoogleMapsModule} from '@angular/google-maps';
 import {HomeComponent} from './home/home.component';
 import {HttpClientModule} from '@angular/common/http';
+import {ItemDialogComponent} from './item-dialog/item-dialog.component';
 import {LoginComponent} from './login/login.component';
+import {MatAutocompleteModule} from '@angular/material/autocomplete';
 import {MatButtonModule} from '@angular/material/button';
 import {MatCardModule} from '@angular/material/card';
 import {MatExpansionModule} from '@angular/material/expansion';
+import {MapComponent} from './map/map.component';
+import {MatDialogModule} from '@angular/material/dialog';
+import {MatDividerModule} from '@angular/material/divider';
 import {MatFormFieldModule} from '@angular/material/form-field';
 import {MatIconModule} from '@angular/material/icon';
 import {MatInputModule} from '@angular/material/input';
@@ -27,28 +37,31 @@ import {MatSelectModule} from '@angular/material/select';
 import {MatSidenavModule} from '@angular/material/sidenav';
 import {MatRadioModule} from '@angular/material/radio';
 import {MatProgressSpinnerModule} from '@angular/material/progress-spinner';
+import {MatSnackBarModule} from '@angular/material/snack-bar';
+import {MatStepperModule} from '@angular/material/stepper';
+import {MatProgressBarModule} from '@angular/material/progress-bar';
 import {MatTabsModule} from '@angular/material/tabs';
 import {MatToolbarModule} from '@angular/material/toolbar';
+import {CurrentProfileTabComponent} from './current-profile-tab/current-profile-tab.component';
+import {NavbarComponent} from './navbar/navbar.component';
+import {NgAisModule} from 'angular-instantsearch';
 import {NgModule} from '@angular/core';
 import {ProfileCardComponent} from './profile-card/profile-card.component';
 import {ProfileMenuComponent} from './profile-menu/profile-menu.component';
-import {NavbarComponent} from './navbar/navbar.component';
 import {RecipePageComponent} from './recipe-page/recipe-page.component';
 import {RecipesComponent} from './recipes/recipes.component';
-import {SetupComponent} from './setup/setup.component';
-import {SignupComponent} from './signup/signup.component';
-import {UploadRecipeComponent} from './upload-recipe/upload-recipe.component'; 
-import {ViewProfilesComponent} from './view-profiles/view-profiles.component'; 
-import {UserPageComponent} from './user-page/user-page.component'; 
-import {environment} from '../environments/environment';
-import {MatAutocompleteModule} from '@angular/material/autocomplete';
-import {MatSlideToggleModule} from '@angular/material/slide-toggle';
-import {NgAisModule} from 'angular-instantsearch';
-import {AutoCompleteComponent} from './auto-complete/auto-complete.component';
 import {SearchBoxComponent} from './search-box/search-box.component';
+import {SetupComponent} from './setup/setup.component';
+import {ShoppingListComponent} from './shopping-list/shopping-list.component';
+import {SignupComponent} from './signup/signup.component';
+import {DiscoverDisplayComponent} from './discover-display/discover-display.component';
+import {RecipeCardComponent} from './recipe-card/recipe-card.component';
+import {ViewProfilesComponent} from './view-profiles/view-profiles.component'; 
+import {UploadRecipeComponent} from './upload-recipe/upload-recipe.component';
+import {MatSlideToggleModule} from '@angular/material/slide-toggle';
 import {DiscoverPageComponent} from './discover-page/discover-page.component';
-import { DiscoverDisplayComponent } from './discover-display/discover-display.component';
-import { AutocompleteComponent } from './autocomplete/autocomplete.component';
+import {AutocompleteComponent} from './autocomplete/autocomplete.component';
+import {UserPageComponent} from './user-page/user-page.component'; 
 
 
 @NgModule({
@@ -63,18 +76,26 @@ import { AutocompleteComponent } from './autocomplete/autocomplete.component';
     SetupComponent,
     HomeComponent,
     ProfileCardComponent,
-    ViewProfilesComponent,
     UserPageComponent,
     UploadRecipeComponent,
+    RecipeCardComponent,
+    MapComponent,
     RecipesComponent,
     RecipePageComponent,
     CurrentProfilePageComponent,
+    ChangeProfileImgComponent,
+    CurrentProfileTabComponent,
+    ViewProfilesComponent,
     SearchBoxComponent,
     DiscoverPageComponent,
     DiscoverDisplayComponent,
     AutocompleteComponent,
+    ShoppingListComponent,
+    ItemDialogComponent,
   ],
-
+  entryComponents: [
+    ItemDialogComponent,
+  ],
   imports: [
     AngularFireModule.initializeApp(environment.firebase),
     AngularFireAuthModule,
@@ -85,31 +106,36 @@ import { AutocompleteComponent } from './autocomplete/autocomplete.component';
     BrowserModule,
     BrowserAnimationsModule,
     FormsModule,
+    GoogleMapsModule,
     HttpClientModule,
-    MatCardModule,
-    MatFormFieldModule,
+    MatAutocompleteModule,
     MatButtonModule,
+    MatCardModule,
+    MatChipsModule,
+    MatDialogModule,
+    MatDividerModule,
     MatExpansionModule,
     MatFormFieldModule,
-    MatInputModule,
-    MatIconModule,
-    MatFormFieldModule,
-    MatAutocompleteModule,
     ReactiveFormsModule,
     MatListModule,
+    MatIconModule,
+    MatInputModule,
     MatMenuModule,
     MatRippleModule,
+    MatProgressBarModule,
     MatProgressSpinnerModule,
     MatRadioModule,
     MatSelectModule,
     MatSidenavModule,
     MatSlideToggleModule,
+    MatStepperModule,
     MatToolbarModule,
-    ReactiveFormsModule,
+    MatSnackBarModule,
     MatTabsModule,
-    NgAisModule.forRoot()
+    MatToolbarModule,
+    NgAisModule.forRoot(),
   ],
-  providers: [],
+  providers: [{provide: LocationStrategy, useClass: HashLocationStrategy}],
   bootstrap: [AppComponent],
 })
 export class AppModule {}
