@@ -1,6 +1,12 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { RecipePageComponent } from './recipe-page.component';
+import {BrowserModule} from '@angular/platform-browser';
+import {BrowserAnimationsModule} from '@angular/platform-browser/animations';
+import {MatTabsModule} from '@angular/material/tabs'
+import {AngularFireModule} from '@angular/fire';
+import {RecipePageComponent} from './recipe-page.component';
+import {environment} from '../../environments/environment';
+import {AngularFirestoreModule} from '@angular/fire/firestore';
+import {RouterTestingModule} from '@angular/router/testing' ;
 
 describe('RecipePageComponent', () => {
   let component: RecipePageComponent;
@@ -8,7 +14,15 @@ describe('RecipePageComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ RecipePageComponent ]
+      declarations: [ RecipePageComponent ],
+      imports: [
+        BrowserAnimationsModule,
+        BrowserModule,
+        MatTabsModule,
+        AngularFireModule.initializeApp(environment.firebase),
+        AngularFirestoreModule,
+        RouterTestingModule
+      ], providers: []
     })
     .compileComponents();
   }));
